refactor: refactor checkout page
